Allow for URL-Encoded token and refresh POST encoding
The purpose of these changes is compatibility with FusionAuth.

These changes add a configuration key of 'tokenPostEncoding' as a pseudo
enumeration that is either 'multi-part' ( the default ) or 'url-encoded'.
Setting this configuration key to a value of 'url-encoded' will cause
the token requests to be sent as 'application/x-www-form-urlencoded'
POSTs. ( FusionAuth does not support multi-part POSTS )

This is accomplished by explictly typing the token request data and
passing that into the funtion performing the fetch. The actual POST
body is then built along with the proper headers in the function
doing the fetching.

This commit also increases the code verifier random length to 96.
( FusionAuth seems to reject shorter code verifiers )