POST encoding to 'application/x-www-form-urlencoded' per OAuth 2.0 spec (#13)
* Always use form url-encoding per OAuth 2.0 spec

The purpose of these changes is compatibility with FusionAuth.

These changes the token requests to be sent as 'application/x-www-form-urlencoded' POSTs. ( FusionAuth does not support multi-part POSTS )

This is accomplished by explicitly typing the token request data and
passing that into the function performing the fetch. The actual POST
body is then built along with the proper headers in the function
doing the fetching.

This commit also increases the code verifier random length to 96.
( FusionAuth seems to reject shorter code verifiers )