fix: handle all 400 refresh requests as "expired"
This commit changes how failed refresh requests are handled.

Since we often have to assume the refreshToken expire time, it will
sometimes be wrong. Therefore, we might make a refresh request with an
expired refreshToken.

This change make sure all 400 status codes on the refresh request are
handled as if the refreshToken has expired. Even if there might be other
reasons.