refactor: make types DRY
Previously, there were some types that were used in multiple places,
but their type definitions were copied/pasted in all of those places.
This was hard to maintain, led to a bug in one instance, and was not
DRY, so this commit defines them in a single place to fix this.