feat: add "replace" login method
Previously, when a login was performed via a redirect, it would always
create a new browser history entry. Sometimes, however, replacing the
current browser history entry is desirable instead, so this commit adds
the ability to specify this as an option to the logIn() function, as
well as an option in the configuration if so desired.

Closes #210